requirements.txt
agno
ollama
openai
groq
python-dotenv
duckduckgo-search
yfinance
lancedb
tantivy
pypdf
requests
exa-py
newspaper4k
lxml_html_clean
sqlalchemy